refactor(frontend): extract shared api config from legacy services
