+++ frontend/src/shared/config/app-config.ts
@@ -1,0 +1,7 @@
+import { getApiUrl } from "./api";
+
+export const appConfig = {
+   apiUrl: getApiUrl(),
+   disableAllAIFunctionality:
+      process.env.DISABLE_ALL_AI_FUNCTIONALITY === "true"
+};
