+++ frontend/src/shared/config/index.ts
@@ -1,1 +1,2 @@
-export {};
+export * from "./api";
+export * from "./app-config";
